+++ package.json
@@ -2,18 +2,21 @@
   "name": "rehabminiapp",
   "version": "1.0.0",
   "private": true,
-  "type": "module",
   "scripts": {
-    "build": "tsc && cp index.html dist/index.html",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
     "test": "echo \"No tests\""
   },
   "dependencies": {
+    "next": "^14.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "typescript": "^5.0.0",
+    "@types/node": "^20.0.0",
     "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.2.0"
+    "@types/react-dom": "^18.2.0",
+    "typescript": "^5.0.0"
   }
 }
